fix: MainScreen 내의 main 함수 main.dart로 분리
