+++ sesac7_flutter_recipe_app/lib/presentation/main/main_screen.dart
@@ -10,25 +10,6 @@ import 'package:flutter_recipe_app/presentation/component/dialog/rating_dialog.d
 import 'package:flutter_recipe_app/presentation/component/tab/tabs.dart';
 import 'package:flutter_recipe_app/presentation/component/text_field/input_filed.dart';
 
-void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-      ),
-      home: const MainScreen(),
-    );
-  }
-}
-
 class MainScreen extends StatefulWidget {
   const MainScreen({super.key});
